+++ clobber/frontend/lib/clobber/internal/parser_error_factory.hpp
@@ -3,8 +3,8 @@
 struct ParserError; // clobber/parser.hpp
 
 namespace ParserErrorFactory {
-ParserError InternalErr(int span_start, int span_len);
-ParserError InternalErr(int error_code, int span_start, int span_len);
+ParserError InternalErr(size_t span_start, size_t span_len);
+ParserError InternalErr(int error_code, size_t span_start, size_t span_len);
 }; // namespace ParserErrorFactory
 
 namespace err = ParserErrorFactory; // alias for convenience
